requirements.txt
fastapi
uvicorn
nltk
pandas
numpy
sklearn
loguru
pprint
absl-py
spacy
uuid
tensorflow
tensorflow_hub
azure-cosmos
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.0/en_core_web_sm-2.2.0.tar.gz